+++ src/Entities/Settings/Course.java
@@ -54,9 +54,9 @@ public class Course implements Serializable {
     @Column(name = "payment_mode")
     private String paymentMode;
     @Column(name = "admission_fee")
-    private Integer admissionFee;
+    private double admissionFee;
     @Column(name = "fee")
-    private Integer fee;
+    private double fee;
     @Column(name = "status")
     private Integer status;
 
@@ -131,19 +131,19 @@ public class Course implements Serializable {
         this.paymentMode = paymentMode;
     }
     
-    public Integer getAdmissionFee() {
+    public double getAdmissionFee() {
         return admissionFee;
     }
 
-    public void setAdmissionFee(Integer admissionFee) {
+    public void setAdmissionFee(double admissionFee) {
         this.admissionFee = admissionFee;
     }
 
-    public Integer getFee() {
+    public double getFee() {
         return fee;
     }
 
-    public void setFee(Integer fee) {
+    public void setFee(double fee) {
         this.fee = fee;
     }
